Able to add sets to particular exercises and display correctly

+++ app/src/main/java/com/example/jim/ui/repositories/SetsRepository.kt
@@ -30,7 +30,7 @@ object SetsRepository {
 
     suspend fun createSet(
         reps: Int,
-        weight: Float,
+        weight: Int,
         exerciseId: String,
         workoutId: String? = "",
     ): Set {
